Split module graph loading tests
